The script-editing and Dub Agent routes live in app/api/script.py and app/api/agent.py; fakes re-pointed, proven by canaries
